requirements.txt
discord.py
beautifulsoup4
httpx
python-dotenv
asyncio
